+++ src/my_context.h
@@ -28,7 +28,7 @@ public:
     void set_camera(Camera* new_camera);
     void switch_camera();
 
-    void update(std::chrono::duration<double> delta) {
+    void update(std::chrono::duration<float> delta) {
         env->dir_light_.set_trans(glm::rotate(env->dir_light_.get_trans(), glm::radians(45.f) * static_cast<float>(delta.count()), glm::vec3(0.f, 1.f, 0.f)));
         env->dir_light_.set_trans(glm::lookAt(env->dir_light_.get_position(), glm::vec3(0.f), glm::vec3(0.f, 0.f, 1.f)));
